Addin the update and the change status

+++ todoapp/views.py
@@ -18,6 +18,29 @@ def index(request):
 	return render(request , 'index.html', {'todos':todos})
 
 def delete(request , todo_id):
+
+	if request.method == 'POST':
+		if request.POST.get('delete'):
+			todo = get_object_or_404(Todo , pk=todo_id)
+			todo.delete()
+			return redirect('index')
+
+		if request.POST.get('complete'):
+			todo = get_object_or_404(Todo , pk=todo_id)
+			if todo.is_complete == True:
+				todo.is_complete = False
+			else:
+				todo.is_complete = True
+				
+			todo.save()
+			return redirect('index')
+
+		if request.POST.get('update'):
+			return render(request , 'update.html' , {'todo_id': todo_id})
+
+
+def update(request , todo_id):
 	todo = get_object_or_404(Todo , pk=todo_id)
-	todo.delete()
+	todo.todo = request.POST['todo']
+	todo.save()
 	return redirect('index')
